[Bronze V] Title: 이상한 기호, Time: 180 ms, Memory: 17700 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,11 +4,11 @@ public class Main {
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
-        int m = sc.nextInt();
+        long n = sc.nextLong();
+        long m = sc.nextLong();
 
-        int plus = (n+m);
-        int minus = (n-m);
+        long plus = (n+m);
+        long minus = (n-m);
         System.out.println(plus * minus);
     }
 }
